+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/film/LikeDao.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.dao.film;
 
 public interface LikeDao {
-    void addLikeToFilm(int id, int userId);
+    void addLikeToFilm(long id, long userId);
 
-    void deleteLikeFromFilm(int id, int userId);
+    void deleteLikeFromFilm(long id, long userId);
 }
