+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/user/FriendDao.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage.dao.user;
 
 public interface FriendDao {
-    void addFriend(int id, int friendId);
+    void addFriend(long id, long friendId);
 
-    void deleteFriend(int id, int friendId);
+    void deleteFriend(long id, long friendId);
 }
